analytics-service/requirements.txt
fastapi>=0.95.0,<0.111.0
uvicorn[standard]>=0.18.0,<0.28.0
sqlalchemy[asyncio]>=2.0.0,<2.1.0
alembic>=1.9.0,<1.14.0
psycopg[binary,pool]>=3.1.0,<3.2.0
pydantic>=2.0.0,<2.7.0
python-dotenv>=0.20.0,<1.1.0

# Celery and RabbitMQ
celery>=5.3.0,<5.4.0
redis>=4.5.0,<5.1.0
aio-pika>=9.0.0,<9.5.0

# JWT Authentication
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0

# Optional: ClickHouse
# clickhouse-connect>=0.5.0,<0.8.0

# Development and Testing
pytest>=7.0.0,<8.2.0
httpx>=0.24.0,<0.28.0 